rip to fmtlog.h
